requirements.txt
cmake
debugpy
dlib
face-recognition
golden-eagle
imutils
japanize-matplotlib
matplotlib
numpy
opencv-python
opencv-contrib-python
pydantic
python-dotenv
pytk
Pillow
setuptools<81.0
wheel
